Added Switch Blueprint Automation
The original automation blueprint from jesserockz was designed for
lights. However, all of my lights are actually on the switch domain. I
modified the blueprint to allow selection of switches instead, and
removed the dimming functionality for now. Ideally, dimming could be
added for separate dimming entities. And also adding functionality for
the un-used buttons if not dimming.